requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.0.5
djangorestframework==3.11.1
djongo==1.3.3
dnspython==2.0.0
pycodestyle==2.6.0
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
toml==0.10.1
